requirements.txt
camel-ai
neo4j
requests
# requirements.txt
gradio
python-dotenv
camel-toolkit
openai
py2neo
qdrant-client
autopep8
# any other libs your docker_agent.py needs… 
docker 

RUN --mount=type=cache,target=/root/.cache/pip \
    pip install --upgrade pip \
 && pip install --no-cache-dir -r requirements.txt 